# internal/http/app/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_abdulazizax_udevslab-lesson3_internal_models.AggregatedOrdersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.OrderAggregate'
        type: array
    type: object
  github_com_abdulazizax_udevslab-lesson3_internal_models.Error:
    properties:
      message:
        type: string
    type: object
  github_com_abdulazizax_udevslab-lesson3_internal_models.Order:
    properties:
      createdAt:
        type: integer
      id:
        type: string
      productId:
        type: string
      quantity:
        type: integer
      status:
        type: string
      total:
        type: number
      updatedAt:
        type: integer
      userId:
        type: string
    type: object
  github_com_abdulazizax_udevslab-lesson3_internal_models.OrderAggregate:
    properties:
      _id:
        properties:
          month:
            type: integer
          year:
            type: integer
        type: object
      orderCount:
        type: integer
      totalAmount:
        type: number
    type: object
  github_com_abdulazizax_udevslab-lesson3_internal_models.OrderCreate:
    properties:
      productId:
        type: string
      quantity:
        type: integer
      status:
        type: string
      userId:
        type: string
    type: object
  github_com_abdulazizax_udevslab-lesson3_internal_models.OrderUpdate:
    properties:
      productId:
        type: string
      quantity:
        type: integer
      status:
        type: string
      userId:
        type: string
    type: object
  github_com_abdulazizax_udevslab-lesson3_internal_models.Pagination:
    properties:
      limit:
        description: 'Query: ?limit=10'
        type: integer
      page:
        description: 'Query: ?page=1'
        type: integer
    required:
    - limit
    - page
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Endpoints for MiniTwitter
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: '# UdevsLab Homework3'
  version: 1.03.67.83.145
paths:
  /orders:
    get:
      description: Retrieve a list of all orders in the database with pagination
      parameters:
      - description: Page information
        in: body
        name: pagination
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Pagination'
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            items:
              $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Order'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
      summary: List all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order and return the created order's ID
      parameters:
      - description: Order information
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.OrderCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Order ID
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
      summary: Create a new order
      tags:
      - Orders
  /orders/{order_id}:
    delete:
      description: Delete an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
        "404":
          description: Order Not Found
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
      summary: Delete an order
      tags:
      - Orders
    get:
      description: Fetch a single order by its ID
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
        "404":
          description: Order Not Found
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
      summary: Get an order by ID
      tags:
      - Orders
    put:
      consumes:
      - application/json
      description: Update the details of an existing order
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      - description: Order fields to update
        in: body
        name: updates
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.OrderUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Order updated successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
        "404":
          description: Order Not Found
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
      summary: Update an order
      tags:
      - Orders
  /orders/aggregates:
    get:
      consumes:
      - application/json
      description: Fetch aggregated orders based on the date range, pagination, and
        sorting (by date).
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: 1
        description: Order (1 for ascending, -1 for descending)
        in: query
        name: order
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Aggregated Orders
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.AggregatedOrdersResponse'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
      summary: Get aggregated orders by date range (year and month)
      tags:
      - Orders
  /orders/customer/{customer_id}:
    get:
      consumes:
      - application/json
      description: Get a list of orders for a specific customer by customer ID
      parameters:
      - description: Customer ID
        in: path
        name: customer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            items:
              $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Order'
            type: array
        "400":
          description: Invalid customer ID
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
        "404":
          description: No orders found
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
      summary: Get orders by customer
      tags:
      - Orders
  /orders/range:
    get:
      description: Retrieve a list of all orders within a date range in the database,
        sorted by created_at in ascending order
      parameters:
      - description: 'Order (-1: decreasing, 1: increasing)'
        in: query
        name: order
        required: true
        type: integer
      - description: Page information
        in: body
        name: pagination
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Pagination'
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            items:
              items:
                $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Order'
              type: array
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_abdulazizax_udevslab-lesson3_internal_models.Error'
      summary: List all orders within a specific date range, with pagination, and
        sorted by date (ascending)
      tags:
      - orders
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"